Move Tab and Enter to right thumb
Tab is very common key for me so it is good to have it on default thumb position. Enter is less frequent and also more destructive for accidental press, so moving it next to default position. Now with only 2 action keys per thumb I expect less cognitive load and faster muscle memory build.